pyproject.toml
[project]
name = "ha-govee-led-ble"
version = "1.8.0"
description = "Home Assistant custom integration for Govee BLE LED lights"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "bleak>=0.21.0",
    "bleak-retry-connector>=3.0.0",
]

[dependency-groups]
dev = [
    "aiousbwatcher>=1.1.1",
    "coverage>=7.0",
    "mypy>=1.15",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-homeassistant-custom-component>=0.13",
    "pyserial>=3.5",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
fail_under = 90

[tool.mypy]
python_version = "3.13"
strict = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "bleak",
    "bleak.*",
    "bleak_retry_connector.*",
    "homeassistant.*",
    "pytest_homeassistant_custom_component.*",
]
ignore_missing_imports = true
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
warn_return_any = false
disable_error_code = ["arg-type", "attr-defined", "no-untyped-call", "typeddict-item"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["custom_components/ha_govee_led_ble/manifest.json:version"]
build_command = ""
commit_message = "chore(release): {version}"
tag_format = "v{version}"

[tool.semantic_release.commit_parser_options]
patch_tags = ["fix", "perf", "build"]
other_allowed_tags = ["chore", "ci", "docs", "style", "refactor", "test"]

[tool.semantic_release.branches.master]
match = "master"
prerelease = false

[tool.semantic_release.changelog]
mode = "init"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.remote]
type = "github"
token = { env = "GITHUB_TOKEN" }
